+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed [here](https://github.com/sensu-plugins/community/blob/master/HOW_WE_CHANGELOG.md)
 
 ## [Unreleased]
+### Fixed
+- check-strongswan-connection.rb: add unknown status if unable to retrieve connection status (@nishiki)
 
 ## [1.1.0] - 2018-02-22
 ### Added